// src/interfaces/Car.ts
export interface ICar {
  _id?: string | FormDataEntryValue | null;
  make: string | FormDataEntryValue | null;
  model: string | FormDataEntryValue | null;
  equipment: string | FormDataEntryValue | null;
  color: string | FormDataEntryValue | null;
  year: string | FormDataEntryValue | null;
  category: string | FormDataEntryValue | null;
  mileage: string | FormDataEntryValue | null;
  price: string | FormDataEntryValue | null;
}